+++ stdlib/auth/src/test/java/org/ballerinalang/stdlib/auth/AuthnHandlerTest.java
@@ -16,7 +16,7 @@
  *  under the License.
  */
 
-package org.ballerinalang.test.auth;
+package org.ballerinalang.stdlib.auth;
 
 import org.ballerinalang.config.ConfigRegistry;
 import org.ballerinalang.launcher.util.BCompileUtil;
